requirements.txt
matplotlib==2.0.0
numpy==1.12.1
scipy==0.19.0
torch==0.4.1
torchvision==0.2.0
tqdm==4.11.2
pydensecrf
protobuf
tensorboardX
pyyaml
pretrainedmodels
opencv-python
